+++ src/Map.jsx
@@ -1,7 +1,6 @@
 import React, { useEffect, useRef, useState } from 'react';
-import PropTypes from 'prop-types';
 import { MapContainer, TileLayer } from 'react-leaflet';
-import { Box, Text } from 'grommet';
+import { Box } from 'grommet';
 import { Grommet } from 'grommet-icons';
 
 import GrommetMarker from './GrommetMarker';
@@ -10,46 +9,10 @@ import { GrommetGeoJSON } from './GrommetGeoJSON';
 import { Pin } from './markers';
 import { generateLocations } from './utils/locations';
 import { geojsonLocations } from './utils/geojson_data';
-import { getStatusCounts, STATUS_MAP } from './utils/status';
+import { ClusterPopup } from './ClusterPopup';
 
 const initialStyle = {
   height: '100%',
-};
-
-const ClusterTip = ({ cluster }) => {
-  const statusCounts = getStatusCounts(cluster.getAllChildMarkers());
-
-  return (
-    <Box gap="xsmall">
-      <Text color="text-strong" weight={500}>
-        Device Summary
-      </Text>
-      {/* TO DO should we not use bold? */}
-      <Text weight="bold" size="xsmall">
-        Location:{' '}
-        <Text size="xsmall" weight="normal">
-          San Francisco, CA
-        </Text>
-      </Text>
-      {Object.entries(statusCounts).map(([key, value]) => {
-        const StatusIcon = STATUS_MAP[key].icon;
-        return value > 0 ? (
-          <Box key={key} align="center" direction="row" gap="xsmall">
-            <StatusIcon color={STATUS_MAP[key].color} size="small" />
-            <Text color="text-strong" size="xsmall">
-              {value} {STATUS_MAP[key].state}
-            </Text>
-          </Box>
-        ) : null;
-      })}
-    </Box>
-  );
-};
-
-ClusterTip.propTypes = {
-  cluster: PropTypes.shape({
-    getAllChildMarkers: PropTypes.func,
-  }),
 };
 
 function Map() {
@@ -94,7 +57,9 @@ function Map() {
             url="https://tiles.stadiamaps.com/tiles/alidade_smooth/{z}/{x}/{y}{r}.png"
           />
           <GrommetMarker position={geolocation} icon={<Grommet />} />
-          <MarkerClusterGroup tip={cluster => <ClusterTip cluster={cluster} />}>
+          <MarkerClusterGroup
+            popup={cluster => <ClusterPopup cluster={cluster} />}
+          >
             {generateLocations(50, { center: geolocation, radius: 5 }).map(
               (location, index) => (
                 <GrommetMarker
@@ -105,7 +70,9 @@ function Map() {
               ),
             )}
           </MarkerClusterGroup>
-          <MarkerClusterGroup tip={cluster => <ClusterTip cluster={cluster} />}>
+          <MarkerClusterGroup
+            popup={cluster => <ClusterPopup cluster={cluster} />}
+          >
             <GrommetGeoJSON key="my-geojson" data={geojsonLocations} />
           </MarkerClusterGroup>
         </MapContainer>
